Fix: ERR | WorkspaceTest | testCreateWorkspace (#1407) 1. Refactored navigateToWorkspace to enable method chaining by returning the HomePage object. 2. Updated WorkspaceTest to utilize the improved navigateToWorkspace method for seamless navigation and chaining. 3. Verified and executed all related test cases to ensure functionality remains intact.
